requirements.txt
# Auto Test Dependencies
Appium-Python-Client>=3.1.0
pytest>=7.4.0
pytest-html>=4.0.0
PyYAML>=6.0
selenium>=4.15.0
flask>=3.0.0
werkzeug>=3.0.0
openpyxl>=3.1.0
